All BioCarta pathways

All BioCarta pathways

all_pathways()

Value

A vector of pathway IDs (the primary pathway IDs on BioCarta).

Details

The original BioCarta website (biocarta.com) is retired, but the full list of pathways can be found from archived websites such as https://web.archive.org/web/20170122225118/https://cgap.nci.nih.gov/Pathways/BioCarta_Pathways or https://www.gsea-msigdb.org/gsea/msigdb/human/genesets.jsp?collection=CP:BIOCARTA.

Examples

all_pathways()
#>   [1] "h_RELAPathway"           "h_no1Pathway"           
#>   [3] "h_gsPathway"             "h_CSKPathway"           
#>   [5] "h_pkcPathway"            "h_srcRPTPPathway"       
#>   [7] "h_amiPathway"            "h_granulocytesPathway"  
#>   [9] "h_lympathway"            "h_lymphocytePathway"    
#>  [11] "h_arapPathway"           "h_agrPathway"           
#>  [13] "h_acrPathway"            "h_akap95Pathway"        
#>  [15] "h_aktPathway"            "h_alkPathway"           
#>  [17] "h_ParkinsonsPathway"     "h_alternativePathway"   
#>  [19] "h_At1rPathway"           "h_ace2Pathway"          
#>  [21] "h_anthraxPathway"        "h_asbcellPathway"       
#>  [23] "h_mhcPathway"            "h_DNAfragmentPathway"   
#>  [25] "h_chemicalPathway"       "h_sppaPathway"          
#>  [27] "h_atmPathway"            "h_agpcrPathway"         
#>  [29] "h_bcrmolecule"           "h_bcellsurvivalPathway" 
#>  [31] "h_blymphocytePathway"    "h_pparPathway"          
#>  [33] "h_sumoPathway"           "h_bcrPathway"           
#>  [35] "h_biopeptidesPathway"    "h_botulinPathway"       
#>  [37] "h_ranklPathway"          "h_bard1Pathway"         
#>  [39] "h_btg2Pathway"           "h_bbcellPathway"        
#>  [41] "h_CaCaMPathway"          "h_cdMacPathway"         
#>  [43] "h_HBxPathway"            "h_flumazenilPathway"    
#>  [45] "h_carm-erPathway"        "h_caspasePathway"       
#>  [47] "h_argininecPathway"      "h_methioninepathway"    
#>  [49] "h_cblPathway"            "h_CCR3Pathway"          
#>  [51] "h_cd40Pathway"           "h_cdc25Pathway"         
#>  [53] "h_mcmPathway"            "h_g1Pathway"            
#>  [55] "h_g2Pathway"             "h_cell2cellPathway"     
#>  [57] "h_LairPathway"           "h_ceramidePathway"      
#>  [59] "h_tidPathway"            "h_chrebpPathway"        
#>  [61] "h_hSWI-SNFpathway"       "h_circadianPathway"     
#>  [63] "h_classicPathway"        "h_compPathway"          
#>  [65] "h_vdrPathway"            "h_hdacPathway"          
#>  [67] "h_gcrPathway"            "h_ctcfPathway"          
#>  [69] "h_ctlPathway"            "h_cxcr4Pathway"         
#>  [71] "h_fbw7Pathway"           "h_ranPathway"           
#>  [73] "h_cellcyclePathway"      "h_cftrPathway"          
#>  [75] "h_cytokinePathway"       "h_inflamPathway"        
#>  [77] "h_d4gdiPathway"          "h_rarPathway"           
#>  [79] "h_dcPathway"             "h_p35alzheimersPathway" 
#>  [81] "h_dicerPathway"          "h_rnaPathway"           
#>  [83] "h_mta3Pathway"           "h_skp2e2fPathway"       
#>  [85] "h_calcineurinPathway"    "h_egfPathway"           
#>  [87] "h_eicosanoidPathway"     "h_etcPathway"           
#>  [89] "h_ndkDynaminPathway"     "h_ephA4Pathway"         
#>  [91] "h_epoPathway"            "h_ecmPathway"           
#>  [93] "h_erkPathway"            "h_erythPathway"         
#>  [95] "h_eponfkbPathway"        "h_eradPathway"          
#>  [97] "h_EfpPathway"            "h_eifPathway"           
#>  [99] "h_extrinsicPathway"      "h_fasPathway"           
#> [101] "h_fcer1Pathway"          "h_fibrinolysisPathway"  
#> [103] "h_fMLPpathway"           "h_fosbPathway"          
#> [105] "h_freePathway"           "h_slrp2Pathway"         
#> [107] "h_fxrPathway"            "h_tubbyPathway"         
#> [109] "h_erbB4pathway"          "h_gabaPathway"          
#> [111] "h_GATA3pathway"          "h_appPathway"           
#> [113] "h_ghrelinPathway"        "h_glycolysisPathway"    
#> [115] "h_setPathway"            "h_ghPathway"            
#> [117] "h_ahspPathway"           "h_tcapoptosisPathway"   
#> [119] "h_vifPathway"            "h_HivnefPathway"        
#> [121] "h_hopPathway"            "h_salmonellaPathway"    
#> [123] "h_mPRPathway"            "h_hcmvPathway"          
#> [125] "h_p53hypoxiaPathway"     "h_hifPathway"           
#> [127] "h_ifnaPathway"           "h_ifngPathway"          
#> [129] "h_igf1Pathway"           "h_il17Pathway"          
#> [131] "h_il18Pathway"           "h_il2Pathway"           
#> [133] "h_il3Pathway"            "h_il4Pathway"           
#> [135] "h_il5Pathway"            "h_il6Pathway"           
#> [137] "h_il10Pathway"           "h_il2rbPathway"         
#> [139] "h_il7Pathway"            "h_IL12Pathway"          
#> [141] "h_il22bppathway"         "h_gsk3Pathway"          
#> [143] "h_deathPathway"          "h_RacCycDPathway"       
#> [145] "h_gleevecpathway"        "h_HuntingtonPathway"    
#> [147] "h_reckPathway"           "h_insulinPathway"       
#> [149] "h_integrinPathway"       "h_iresPathway"          
#> [151] "h_intrinsicPathway"      "h_ionPathway"           
#> [153] "h_raccPathway"           "h_keratinocytePathway"  
#> [155] "h_tcraPathway"           "h_lectinPathway"        
#> [157] "h_g1pPathway"            "h_pyk2Pathway"          
#> [159] "h_Lis1Pathway"           "h_LDLpathway"           
#> [161] "h_malatePathway"         "h_egfr_smrtePathway"    
#> [163] "h_mapkPathway"           "h_mCalpainPathway"      
#> [165] "h_AcetaminophenPathway"  "h_pparaPathway"         
#> [167] "h_npcPathway"            "h_melanocytepathway"    
#> [169] "h_cb1rPathway"           "h_etsPathway"           
#> [171] "h_cptPathway"            "h_monocytePathway"      
#> [173] "h_mspPathway"            "h_mTORPathway"          
#> [175] "h_mrpPathway"            "h_pitx2Pathway"         
#> [177] "h_igf1rPathway"          "h_ngfPathway"           
#> [179] "h_vipPathway"            "h_ErbB3Pathway"         
#> [181] "h_neutrophilPathway"     "h_nfkbPathway"          
#> [183] "h_nfatPathway"           "h_nthiPathway"          
#> [185] "h_nos1Pathway"           "h_no2il12Pathway"       
#> [187] "h_rarrxrPathway"         "h_nuclearRsPathway"     
#> [189] "h_aifPathway"            "h_tertpathway"          
#> [191] "h_tercPathway"           "h_ox40Pathway"          
#> [193] "h_arenrf2Pathway"        "h_p38mapkPathway"       
#> [195] "h_p53Pathway"            "h_pdgfPathway"          
#> [197] "h_pelp1Pathway"          "h_Ccr5Pathway"          
#> [199] "h_ptdinsPathway"         "h_plcdPathway"          
#> [201] "h_plcPathway"            "h_plcePathway"          
#> [203] "h_edg1Pathway"           "h_cdk5Pathway"          
#> [205] "h_myosinPathway"         "h_plateletAppPathway"   
#> [207] "h_cpsfPathway"           "h_ps1Pathway"           
#> [209] "h_prionPathway"          "h_pepiPathway"          
#> [211] "h_akapCentrosomePathway" "h_notchpathway"         
#> [213] "h_ptenPathway"           "h_rabPathway"           
#> [215] "h_rac1Pathway"           "h_rasPathway"           
#> [217] "h_nkcellsPathway"        "h_rbPathway"            
#> [219] "h_reelinPathway"         "h_badPathway"           
#> [221] "h_plk3Pathway"           "h_ck1Pathway"           
#> [223] "h_eif2Pathway"           "h_eif4Pathway"          
#> [225] "h_stemPathway"           "h_dspPathway"           
#> [227] "h_p27Pathway"            "h_pgc1aPathway"         
#> [229] "h_cremPathway"           "h_sam68Pathway"         
#> [231] "h_pmlPathway"            "h_npp1Pathway"          
#> [233] "h_dreampathway"          "h_leptinPathway"        
#> [235] "h_rhoPathway"            "h_akap13Pathway"        
#> [237] "h_RNApol3Pathway"        "h_antisensePathway"     
#> [239] "h_atrbrcaPathway"        "h_cardiacegfPathway"    
#> [241] "h_her2Pathway"           "h_erk5Pathway"          
#> [243] "h_malPathway"            "h_mef2dPathway"         
#> [245] "h_mitochondriaPathway"   "h_achPathway"           
#> [247] "h_parkinPathway"         "h_cdc42racPathway"      
#> [249] "h_ppargPathway"          "h_ranMSpathway"         
#> [251] "h_tob1Pathway"           "h_barr-mapkPathway"     
#> [253] "h_bArrestin-srcPathway"  "h_SARSpathway"          
#> [255] "h_hesPathway"            "h_nktPathway"           
#> [257] "h_malatexPathway"        "h_MITRPathway"          
#> [259] "h_il1rPathway"           "h_metPathway"           
#> [261] "h_gpcrPathway"           "h_igf1mtorpathway"      
#> [263] "h_slrpPathway"           "h_soddPathway"          
#> [265] "h_shhPathway"            "h_ptc1Pathway"          
#> [267] "h_smPathway"             "h_spryPathway"          
#> [269] "h_s1pPathway"            "h_stat3Pathway"         
#> [271] "h_stathminPathway"       "h_aMANpathway"          
#> [273] "h_hsp27Pathway"          "h_ctbp1Pathway"         
#> [275] "h_ranbp2Pathway"         "h_PDZsPathway"          
#> [277] "h_tcrmolecule"           "h_tcrPathway"           
#> [279] "h_tcytotoxicPathway"     "h_thelperPathway"       
#> [281] "h_tall1Pathway"          "h_telPathway"           
#> [283] "h_tgfbPathway"           "h_th1th2Pathway"        
#> [285] "h_41bbPathway"           "h_ctla4Pathway"         
#> [287] "h_longevityPathway"      "h_pcafpathway"          
#> [289] "h_prc2Pathway"           "h_eosinophilsPathway"   
#> [291] "h_eea1Pathway"           "h_Par1Pathway"          
#> [293] "h_stressPathway"         "h_tnfr1Pathway"         
#> [295] "h_tnfr2Pathway"          "h_tollPathway"          
#> [297] "h_TPOPathway"            "h_crebPathway"          
#> [299] "h_carm1Pathway"          "h_dbpbPathway"          
#> [301] "h_tffPathway"            "h_trkaPathway"          
#> [303] "h_tsp1Pathway"           "h_arfPathway"           
#> [305] "h_ucalpainPathway"       "h_vegfPathway"          
#> [307] "h_vobesityPathway"       "h_rhodopsinPathway"     
#> [309] "h_vitCBPathway"          "h_WNVpathway"           
#> [311] "h_wntPathway"            "h_wnt-lrp6Pathway"      
#> [313] "h_actinYPathway"         "h_bArrestinPathway"